The routing finally works without known bugs

+++ src/osmCore/CWaySegment.h
@@ -12,7 +12,7 @@ public:
   void UpdateGeometry();
 
   inline double getLength() const { return m_length;};
-  inline double getHeading() const { return m_heading;};
+  inline int getHeading() const { return m_heading;};
 
 
   tOSMNodeShPtr getBeginNode() const;
@@ -21,7 +21,7 @@ public:
 private:
   double m_length;
 
-  double m_heading;
+  int m_heading;
 
   tOSMNodeShPtr m_beginNode;
   tOSMNodeShPtr m_endNode;
